unit tests + removed duplicated code
